Move all components up one level and update all paths to it
