+++ src/__tests__/Footer.test.tsx
@@ -1,5 +1,5 @@
 import { render } from '@testing-library/react';
-import Footer from '../components/Template/Footer';
+import Footer from '../components/Footer';
 
 describe('Footer', () => {
   it('should render without crashing', () => {
